requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
